Create a Cargo workspace
