+++ matrix/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "bit-matrix"
-version = "0.9.0"
-
+version.workspace = true
+rust-version.workspace = true
 authors = [ "Piotr Czarnecki <pioczarn@gmail.com>" ]
 description = "Library for bit matrices and vectors."
 keywords = ["container", "bit", "bitfield", "algebra"]
@@ -9,7 +9,6 @@ documentation = "https://docs.rs/bit-matrix/latest/bit_matrix/"
 repository = "https://github.com/pczarn/bit-matrix"
 license = "MIT/Apache-2.0"
 edition = "2021"
-rust-version = "1.77"
 
 [lib]
 name = "bit_matrix"
