+++ set/Cargo.toml
@@ -1,6 +1,7 @@
 [package]
 name = "bit-set"
-version = "0.8.0"
+version.workspace = true
+rust-version.workspace = true
 authors = ["Alexis Beingessner <a.beingessner@gmail.com>"]
 license = "Apache-2.0 OR MIT"
 description = "A set of bits"
@@ -10,13 +11,12 @@ documentation = "https://docs.rs/bit-set/"
 keywords = ["data-structures", "bitset"]
 readme = "README.md"
 edition = "2021"
-rust-version = "1.85"
 
 [dependencies]
-borsh = { version = "1.5", default-features = false, features = ["derive"], optional = true }
-serde = { version = "1.0", default-features = false, features = ["derive"], optional = true }
+borsh = { version = "1.6.0", default-features = false, features = ["derive"], optional = true }
+serde = { version = "1.0", default-features = false, features = ["derive", "alloc"], optional = true }
 miniserde = { version = "0.1", optional = true }
-nanoserde = { git = "https://github.com/not-fl3/nanoserde.git", optional = true }
+nanoserde = { version = "0.1", optional = true }
 smallvec = { version = "1.15", optional = true }
 
 [dependencies.bit-vec]
@@ -38,9 +38,5 @@ serde = ["dep:serde", "bit-vec/serde"]
 miniserde = ["dep:miniserde", "bit-vec/miniserde"]
 nanoserde = ["dep:nanoserde", "bit-vec/nanoserde"]
 
-serde_std = ["std", "serde/std"]
-serde_no_std = ["serde/alloc"]
-borsh_std = ["borsh/std"]
-
 [package.metadata.docs.rs]
-features = ["borsh", "serde", "miniserde", "nanoserde"]
+features = ["borsh", "serde", "miniserde", "nanoserde", "smallvec"]
