+++ vec/Cargo.toml
@@ -1,6 +1,7 @@
 [package]
 name = "bit-vec"
-version = "0.9.0"
+version.workspace = true
+rust-version.workspace = true
 authors = ["Alexis Beingessner <a.beingessner@gmail.com>"]
 license = "Apache-2.0 OR MIT"
 description = "A vector of bits"
@@ -10,13 +11,12 @@ documentation = "https://docs.rs/bit-vec/"
 keywords = ["data-structures", "bitvec", "bitmask", "bitmap", "bit"]
 readme = "README.md"
 edition = "2021"
-rust-version = "1.85"
 
 [dependencies]
-borsh = { version = "1.5.7", default-features = false, features = ["derive"], optional = true }
-serde = { version = "1.0", default-features = false, features = ["derive"], optional = true }
+borsh = { version = "1.6.0", default-features = false, features = ["derive"], optional = true }
+serde = { version = "1.0", default-features = false, features = ["derive", "alloc"], optional = true }
 miniserde = { version = "0.1", optional = true }
-nanoserde = { git = "https://github.com/not-fl3/nanoserde.git", version = "0.2.2", optional = true }
+nanoserde = { version = "0.1", optional = true }
 smallvec = { version = "1.15", optional = true }
 
 [dev-dependencies]
@@ -27,10 +27,7 @@ generic-tests = "0.1"
 
 [features]
 default = ["std"]
-serde_std = ["std", "serde/std"]
-serde_no_std = ["serde/alloc"]
-borsh_std = ["borsh/std"]
-std = []
+std = ["serde?/std", "borsh?/std"]
 allocator_api = []
 
 [package.metadata.docs.rs]
