+++ vec/LICENSE-APACHE
